requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==4.16.0
scikit-learn
dill
mlflow
pyaml
dagshub
fastapi
uvicorn
starlette 
jinja2
python-multipart
# -e .
